+++ app/drawers/messages/read.css
@@ -1,0 +1,1 @@
+fab{background-color:#8190a4}#status-text{text-transform:uppercase;text-align:right;horizontal-align:right}.fab-button{height:50;width:50;margin:15;background-color:#8190a4;horizontal-align:right;vertical-align:bottom}
