+++ app/shared/cothority/lib/identity.js
