+++ app/shared/cothority/lib/index.js
