+++ app/shared/cothority/lib/misc/misc.js
