+++ app/shared/cothority/lib/omniledger/Argument.js
