+++ app/shared/cothority/lib/omniledger/ClientTransaction.js
