+++ app/shared/cothority/lib/omniledger/Config.js
