+++ app/shared/cothority/lib/omniledger/Delete.js
