+++ app/shared/cothority/lib/omniledger/Instance.js
