+++ app/shared/cothority/lib/omniledger/Instruction.js
