+++ app/shared/cothority/lib/omniledger/OmniledgerRPC.js
