+++ app/shared/cothority/lib/omniledger/Proof.js
