+++ app/shared/cothority/lib/omniledger/Spawn.js
