+++ app/shared/cothority/lib/omniledger/contracts/PopPartyInstance.js
