+++ app/shared/cothority/lib/omniledger/contracts/index.js
