+++ app/shared/cothority/lib/omniledger/darc/Identity.js
