+++ app/shared/cothority/lib/omniledger/darc/IdentityEd25519.js
