+++ app/shared/cothority/lib/omniledger/darc/Request.js
