+++ app/shared/cothority/lib/omniledger/darc/Signature.js
