+++ app/shared/cothority/lib/omniledger/darc/Signer.js
