+++ app/shared/cothority/lib/omniledger/darc/SignerEd25519.js
