+++ app/shared/cothority/lib/omniledger/darc/index.js
