+++ app/shared/cothority/lib/shims/ws.js
