+++ app/shared/cothority/lib/skipchain/index.js
