+++ app/shared/cothority/lib/skipchain/skipchain.js
